requirements.txt
tensorflow-cpu==2.13.0
gradio==4.44.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0
